% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_birth_by_year.R
\docType{data}
\name{birth_by_year}
\alias{birth_by_year}
\title{birth_by_year}
\format{
A data frame with 121 rows and 2 variables:
\describe{
  \item{ year }{  integer }
  \item{ n }{  integer }
}
}
\source{
Source
}
\usage{
birth_by_year
}
\description{
Description.
}
\keyword{datasets}
