% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_birth_by_year_and_sex.R
\docType{data}
\name{birth_by_year_and_sex}
\alias{birth_by_year_and_sex}
\title{birth_by_year_and_sex}
\format{
A data frame with 242 rows and 3 variables:
\describe{
  \item{ year }{  integer }
  \item{ sex }{  character }
  \item{ n }{  integer }
}
}
\source{
Source
}
\usage{
birth_by_year_and_sex
}
\description{
Description.
}
\keyword{datasets}
