% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_colin.R
\docType{data}
\name{colin}
\alias{colin}
\title{colin}
\format{
A data frame with 754 rows and 6 variables:
\describe{
  \item{ year }{  integer }
  \item{ sex }{  character }
  \item{ name }{  character }
  \item{ n }{  integer }
  \item{ dpt }{  character }
  \item{ prop }{  numeric }
}
}
\source{
Source
}
\usage{
colin
}
\description{
Description.
}
\keyword{datasets}
