% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_unique_names.R
\docType{data}
\name{unique_names}
\alias{unique_names}
\title{unique_names}
\format{
A data frame with  rows and  variables:
\describe{
}
}
\source{
Source
}
\usage{
unique_names
}
\description{
Description.
}
\keyword{datasets}
